feat(custom): 实现pk连线功能，完善PK送礼消费者方处理逻辑

+++ live-framework/live-framework-redis-starter/src/main/java/top/flobby/live/framework/redis/starter/key/GiftProviderCacheKeyBuilder.java
@@ -33,11 +33,11 @@ public class GiftProviderCacheKeyBuilder extends RedisKeyBuilder {
                 super.getSplitItem() + uuid;
     }
 
-    public String buildLivingPkKey(Long roomId) {
+    public String buildLivingPkKey(Integer roomId) {
         return super.getPrefix() + LIVING_PK_KEY + super.getSplitItem() + roomId;
     }
 
-    public String buildLivingPkSendSeq(Long roomId) {
+    public String buildLivingPkSendSeq(Integer roomId) {
         return super.getPrefix() + LIVING_PK_SEND_SEQ + super.getSplitItem() + roomId;
     }
 }
